# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new product",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term matched against product name and description",
            "schema": {
              "example": "keyboard",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter products by category id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "One-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (max 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProductsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a list of products with pagination and filtering",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get a single product by ID",
        "tags": [
          "Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Marketplace API",
    "description": "API documentation for the Marketplace application",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Public product name",
            "example": "Wireless Keyboard"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly unique slug",
            "example": "wireless-keyboard"
          },
          "description": {
            "type": "string",
            "description": "Optional rich product description",
            "example": "Compact Bluetooth keyboard with multi-device support"
          },
          "price": {
            "type": "number",
            "description": "Price as a decimal number with two fractional digits",
            "example": 49.9,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "description": "Current available stock quantity",
            "example": 120,
            "minimum": 0
          },
          "imageUrl": {
            "type": "string",
            "description": "Optional image URL shown in clients",
            "example": "https://cdn.example.com/products/keyboard.jpg"
          },
          "categoryId": {
            "type": "string",
            "description": "Target category identifier",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "default": true,
            "description": "Whether this product is publicly visible for purchase"
          }
        },
        "required": [
          "name",
          "slug",
          "price",
          "stock",
          "categoryId"
        ]
      },
      "ProductCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Electronics"
          },
          "slug": {
            "type": "string",
            "example": "electronics"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Wireless Keyboard"
          },
          "slug": {
            "type": "string",
            "example": "wireless-keyboard"
          },
          "description": {
            "type": "string",
            "example": "Compact Bluetooth keyboard"
          },
          "price": {
            "type": "string",
            "description": "Product price as a decimal string with two fraction digits",
            "example": "49.90"
          },
          "stock": {
            "type": "number",
            "example": 120
          },
          "imageUrl": {
            "type": "string",
            "example": "https://cdn.example.com/products/keyboard.jpg"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategoryDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "stock",
          "isActive",
          "categoryId",
          "category",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductPaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 120
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedProductsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ProductPaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}